requirements.txt
fastapi>=0.110.0
uvicorn
sqlalchemy
psycopg2-binary
python-jose
cryptography
redis
bcrypt
requests
celery
python-dotenv
pydantic>=2.6.0
